# server/model_group.go
/*
 * Онлайн библиотека песен
 *
 * тестовое задание
 *
 * API version: 0.0.1
 * Contact: apa.birds@mail.ru
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package server

type Group struct {
	Id   string `json:"id,omitempty"`
	Name string `json:"name,omitempty"`
}